+++ web/web-base/src/main/java/org/visallo/web/routes/vertex/VertexPropertyDetails.java
@@ -13,12 +13,11 @@ import org.visallo.core.security.VisibilityTranslator;
 import org.visallo.core.user.User;
 import org.visallo.core.util.VisalloLogger;
 import org.visallo.core.util.VisalloLoggerFactory;
-import org.visallo.web.BadRequestException;
-import org.visallo.web.VisalloResponse;
 import org.visallo.web.clientapi.model.ClientApiSourceInfo;
 import org.visallo.web.clientapi.model.ClientApiVertexPropertyDetails;
 import org.visallo.web.clientapi.model.VisibilityJson;
 import org.visallo.web.parameterProviders.ActiveWorkspaceId;
+import org.visallo.web.util.VisibilityValidator;
 
 import java.util.ResourceBundle;
 
@@ -48,14 +47,16 @@ public class VertexPropertyDetails implements ParameterizedHandler {
             @ActiveWorkspaceId String workspaceId,
             ResourceBundle resourceBundle,
             User user,
-            Authorizations authorizations,
-            VisalloResponse response
+            Authorizations authorizations
     ) throws Exception {
-        Visibility visibility = visibilityTranslator.toVisibility(visibilitySource).getVisibility();
-        if (!graph.isVisibilityValid(visibility, authorizations)) {
-            LOGGER.warn("%s is not a valid visibility for %s user", visibilitySource, user.getDisplayName());
-            throw new BadRequestException("visibilitySource", resourceBundle.getString("visibility.invalid"));
-        }
+        Visibility visibility = VisibilityValidator.validate(
+                graph,
+                visibilityTranslator,
+                resourceBundle,
+                visibilitySource,
+                user,
+                authorizations
+        );
 
         Vertex vertex = this.graph.getVertex(vertexId, authorizations);
         if (vertex == null) {